+++ common/src/main/java/com/faboslav/variantsandventures/common/init/VariantsAndVenturesEntityTypes.java
@@ -2,6 +2,7 @@ package com.faboslav.variantsandventures.common.init;
 
 import com.faboslav.variantsandventures.common.VariantsAndVentures;
 import com.faboslav.variantsandventures.common.entity.mob.GelidEntity;
+import com.faboslav.variantsandventures.common.entity.mob.MurkEntity;
 import com.faboslav.variantsandventures.common.entity.mob.ThicketEntity;
 import com.faboslav.variantsandventures.common.entity.mob.VerdantEntity;
 import com.faboslav.variantsandventures.common.events.lifecycle.RegisterEntityAttributesEvent;
@@ -24,12 +25,14 @@ public final class VariantsAndVenturesEntityTypes
 	public static boolean previousUseChoiceTypeRegistrations = SharedConstants.useChoiceTypeRegistrations;
 
 	public static final Supplier<EntityType<GelidEntity>> GELID;
+	public static final Supplier<EntityType<MurkEntity>> MURK;
 	public static final Supplier<EntityType<ThicketEntity>> THICKET;
 	public static final Supplier<EntityType<VerdantEntity>> VERDANT;
 
 	static {
 		SharedConstants.useChoiceTypeRegistrations = false;
 		GELID = ENTITY_TYPES.register("gelid", () -> EntityType.Builder.create(GelidEntity::new, SpawnGroup.MONSTER).setDimensions(0.6F, 1.95F).allowSpawningInside(Blocks.POWDER_SNOW).maxTrackingRange(8).build(VariantsAndVentures.makeStringID("gelid")));
+		MURK = ENTITY_TYPES.register("murk", () -> EntityType.Builder.create(MurkEntity::new, SpawnGroup.MONSTER).setDimensions(0.6F, 1.99F).maxTrackingRange(8).build(VariantsAndVentures.makeStringID("murk")));
 		THICKET = ENTITY_TYPES.register("thicket", () -> EntityType.Builder.create(ThicketEntity::new, SpawnGroup.MONSTER).setDimensions(0.6F, 1.95F).maxTrackingRange(8).build(VariantsAndVentures.makeStringID("thicket")));
 		VERDANT = ENTITY_TYPES.register("verdant", () -> EntityType.Builder.create(VerdantEntity::new, SpawnGroup.MONSTER).setDimensions(0.6F, 1.99F).maxTrackingRange(8).build(VariantsAndVentures.makeStringID("verdant")));
 		SharedConstants.useChoiceTypeRegistrations = previousUseChoiceTypeRegistrations;
@@ -37,6 +40,7 @@ public final class VariantsAndVenturesEntityTypes
 
 	public static void registerEntityAttributes(RegisterEntityAttributesEvent event) {
 		event.register(VariantsAndVenturesEntityTypes.GELID.get(), GelidEntity.createGelidAttributes());
+		event.register(VariantsAndVenturesEntityTypes.MURK.get(), MurkEntity.createMurkAttributes());
 		event.register(VariantsAndVenturesEntityTypes.THICKET.get(), ThicketEntity.createZombieAttributes());
 		event.register(VariantsAndVenturesEntityTypes.VERDANT.get(), VerdantEntity.createAbstractSkeletonAttributes());
 	}